arch/avr32/os/context.c: Fix missing cast as reported by avr32-gcc (AVR_32_bit_GNU_Toolchain_3.4.1_348) 4.4.3.
git-svn-id: svn+ssh://svn.code.sf.net/p/ethernut/code/trunk@5436 1ca430b3-8271-4a6b-9b0e-3dcac98a8924

+++ nut/arch/avr32/os/context.c
@@ -205,7 +205,7 @@ HANDLE NutThreadCreate(char *name, void (*fn) (void *), void *arg, size_t stackS
     if ((threadMem = NutHeapAlloc(alloc_size)) == 0) {
         return 0;
     }
-    td = (NUTTHREADINFO *) ((threadMem + stackSize) & ~7);
+    td = (NUTTHREADINFO *) (((size_t)threadMem + stackSize) & ~7);
     ef = (ENTERFRAME *) ((uptr_t) td - sizeof(ENTERFRAME));
     sf = (SWITCHFRAME *) ((uptr_t) ef - sizeof(SWITCHFRAME));
